Moved and consolidated components
