--- src/app/createTrip/[email]/loading.tsx
@@ -1,8 +1,0 @@
-import styles from './page.module.css';
-export default function Loading() {
-    return (
-        <div className={styles.title}>
-            Loading.....
-        </div>
-    );
-}
