+++ src/app/createTrip/[email]/page.tsx
@@ -1,10 +1,10 @@
 'use server';
 import styles from './page.module.css';
-import CreateTripForm from './createTripForm';
+import CreateTripForm from '../../../components/createTrip/createTripForm';
 // import TrailerSectionOfForm from './trailerSectionOfForm';
-import CarrierAndTruckInfo from './carrierAndTruckInfo';
-import TruckCheckBoxesForForm from './truckCheckBoxesForForm';
-import SubmitButton from './submitButton'
+import CarrierAndTruckInfo from '../../../components/createTrip/carrierAndTruckInfo';
+import TruckCheckBoxesForForm from '../../../components/createTrip/truckCheckBoxesForForm';
+import SubmitButton from '../../../components/createTrip/submitButton'
 
 export default async function CreateTrip({ params }: { params: { email: string } }) {
     const decodedEmail = decodeURIComponent(params.email);
