+++ src/components/createTrip/carrierAndTruckInfo.tsx
