+++ src/components/createTrip/trailerSectionOfForm.tsx
