+++ src/components/home/defectStats.tsx
