+++ src/components/home/tripStats.tsx
